requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
asgiref=3.8.1=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
django=4.2.9=pypi_0
djangorestframework=3.14.0=pypi_0
djangorestframework-simplejwt=5.5.0=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
ncurses=6.4=h6a678d5_0
openssl=3.0.16=h5eee18b_0
pip=25.1=pyhc872135_2
pyjwt=2.9.0=pypi_0
python=3.11.11=he870216_0
python-decouple=3.8=pypi_0
python-dotenv=1.1.0=pypi_0
pytz=2025.2=pypi_0
readline=8.2=h5eee18b_0
setuptools=78.1.1=py311h06a4308_0
sqlite=3.45.3=h5eee18b_0
sqlparse=0.5.3=pypi_0
tk=8.6.14=h39e8969_0
tzdata=2025b=h04d1e81_0
wheel=0.45.1=py311h06a4308_0
xz=5.6.4=h5eee18b_1
zlib=1.2.13=h5eee18b_1
